requirements.txt
# Hello-Agents Project Dependencies
# Basic Dependencies
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
jupyter>=1.0.0

# LLM Related
openai>=1.0.0
langchain>=0.0.267
transformers>=4.30.0
tiktoken>=0.4.0

# Agent Frameworks
autogen-agentchat>=0.2.0
agentscope>=0.2.0
hello-agents>=0.2.2
langgraph>=0.0.15

# Memory and Retrieval
faiss-cpu>=1.7.4
chromadb>=0.4.13

# Web Applications
flask>=2.3.0
streamlit>=1.28.0
plotly>=5.15.0

# Tools and Utilities
pydantic>=2.0.0
tqdm>=4.65.0
pytest>=7.3.1
tavily-python>=0.2.2
google-search-results==2.1.0